moved terrain into world package
