+++ core/src/com/jingyuyao/tactical/data/LevelTerrainsLoader.java
@@ -5,8 +5,8 @@ import com.badlogic.gdx.maps.MapProperties;
 import com.badlogic.gdx.maps.tiled.TiledMap;
 import com.badlogic.gdx.maps.tiled.TiledMapTileLayer;
 import com.google.common.base.Preconditions;
-import com.jingyuyao.tactical.model.terrain.Terrain;
 import com.jingyuyao.tactical.model.world.Coordinate;
+import com.jingyuyao.tactical.model.world.Terrain;
 import java.util.HashMap;
 import java.util.Map;
 import javax.inject.Inject;
