+++ core/test/com/jingyuyao/tactical/GameStateTest.java
@@ -17,8 +17,8 @@ import com.jingyuyao.tactical.model.script.Script;
 import com.jingyuyao.tactical.model.ship.Ship;
 import com.jingyuyao.tactical.model.state.Turn;
 import com.jingyuyao.tactical.model.state.WorldState;
-import com.jingyuyao.tactical.model.terrain.Terrain;
 import com.jingyuyao.tactical.model.world.Coordinate;
+import com.jingyuyao.tactical.model.world.Terrain;
 import com.jingyuyao.tactical.model.world.World;
 import java.util.ArrayList;
 import java.util.HashMap;
